+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- feat: add `TransactionFilter<TxType>` to pallet-starknet `Config`
 - chore: remove `ignore` from
   `storage_changes_should_revert_on_transaction_revert` test
 - dev: Implement tests for new rpc method starknet_getTransactionStatus